Crate iterator_markers

Extra marker traits for iterators.

Feature flags:

unstable: Implement for Range and RangeInclusive.

(opt-out) std: Implement for the map and set iterators in std and don't set #[no_std].

Traits

AscendingIterator
Marker trait for iterators that return items in a sorted ascending order. It does not guarantee uniqueness, but equal items must come straight after each other. Use I: UniqueIterator+AscendingIterator if you need both.

DescendingIterator
Marker trait for iterators that return items in a sorted desscending order. It does not guarantee uniqueness, but equal items must come straight after each other. Use I: UniqueIterator+DescendingIterator if you need both.

UniqueIterator
Marker trait for iterators that will never return two equal items, like a set.