Trait iterator_markers::DescendingIterator

pub unsafe trait DescendingIterator: Iterator where Self::Item: PartialOrd { }

Marker trait for iterators that return items in a sorted desscending order. It does not guarantee uniqueness, but equal items must come straight after each other. Use I: UniqueIterator+DescendingIterator if you need both.

The trait is unsafe so consumers can rely on it.
It is a logic error to implement both AscendingIterator and DescendingIterator for the same type.

Examples:

let _ = (1..10).rev();