Trait iterator_markers::UniqueIterator

pub unsafe trait UniqueIterator: Iterator where Self::Item: PartialEq { }

Marker trait for iterators that will never return two equal items, like a set.

The trait is unsafe so consumers can rely on it.